ensure that the matcher runs on the first `tick`
In the past the matcher was only started on the first tick but
nulceo did not wait for its result, so it always looked like the matcher
timed out.

Nulceo already correctly handles this case in case the pattern changes
or the item list was cleared, so we just had to set `cleared` to `true`
when the `Nucleo` instance is created.