Slightly change score strategy for columns
Previously, if a column had no match, it would short circuit
returning None.

This means that if you have a match term on a
second column, but the first didn't match, you wouldn't get any
match for the row.

This change means that non-matching columns are scored as zero,
and then if the match has zero score after all columns, then we
return zero.